remove unneeded test vectors and move some into tests
