+++ Cargo.toml
@@ -7,24 +7,16 @@ members = [
     "soroban-sdk-macros",
     "soroban-spec",
     "tests/empty",
-    "tests/hello",
-    "tests/add_i32",
-    "tests/add_i64",
     "tests/add_u64",
     "tests/add_bigint",
-    "tests/vec",
     "tests/import_contract",
     "tests/invoke_contract",
-    "tests/call_stack",
     "tests/udt",
     "tests/contract_data",
-    "tests/create_contract",
     "tests/events",
     "tests/logging",
     "tests/errors",
-    "tests/account",
 ]
-
 [patch.crates-io]
 soroban-sdk = { path = "soroban-sdk" }
 soroban-auth = { path = "soroban-auth" }
