--- tests/add_i32/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "example_add_i32"
-version = "0.0.4"
-authors = ["Stellar Development Foundation <info@stellar.org>"]
-license = "Apache-2.0"
-edition = "2021"
-publish = false
-rust-version = "1.64"
-
-[lib]
-crate-type = ["cdylib", "rlib"]
-doctest = false
-
-[dependencies]
-soroban-sdk = {path = "../../soroban-sdk"}
-
-[dev-dependencies]
-soroban-sdk = {path = "../../soroban-sdk", features = ["testutils"]}
-
-[features]
-testutils = ["soroban-sdk/testutils"]
