--- tests/add_i32/src/lib.rs
@@ -1,31 +1,0 @@
-#![no_std]
-use soroban_sdk::contractimpl;
-
-pub struct Contract;
-
-#[contractimpl]
-impl Contract {
-    pub fn add(a: i32, b: i32) -> i32 {
-        a + b
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use soroban_sdk::{BytesN, Env};
-
-    use crate::{Contract, ContractClient};
-
-    #[test]
-    fn test_add() {
-        let e = Env::default();
-        let contract_id = BytesN::from_array(&e, &[0; 32]);
-        e.register_contract(&contract_id, Contract);
-        let client = ContractClient::new(&e, &contract_id);
-
-        let x = 10i32;
-        let y = 12i32;
-        let z = client.add(&x, &y);
-        assert!(z == 22);
-    }
-}
